19-pharma-forecaster/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.backends"

[project]
name = "pharma-forecast"
version = "1.0.0"
description = "Time Series Forecasting for pharmaceutical demand, drug shortages, and adverse event trends"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.11"
authors = [
    {name = "PharmaForecast Team"},
]

dependencies = [
    "scikit-learn>=1.4",
    "statsmodels>=0.14",
    "prophet>=1.1",
    "numpy>=1.26",
    "pandas>=2.2",
    "fastapi>=0.110",
    "uvicorn[standard]>=0.27",
    "boto3>=1.34",
    "pydantic>=2.6",
    "structlog>=24.1",
    "plotly>=5.18",
    "httpx>=0.27",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-cov>=4.1",
    "pytest-asyncio>=0.23",
    "ruff>=0.3",
    "mypy>=1.8",
    "pandas-stubs>=2.2",
    "types-boto3>=1.34",
]
airflow = [
    "apache-airflow>=2.8",
]

[project.scripts]
pharma-forecast = "pharma_forecast.api.main:run_server"

[tool.hatch.build.targets.wheel]
packages = ["src/pharma_forecast"]

[tool.ruff]
target-version = "py311"
line-length = 100
src = ["src"]

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP", "B", "A", "SIM", "TCH"]
ignore = ["E501"]

[tool.ruff.lint.isort]
known-first-party = ["pharma_forecast"]

[tool.mypy]
python_version = "3.11"
strict = true
warn_return_any = true
warn_unused_configs = true

[[tool.mypy.overrides]]
module = [
    "prophet.*",
    "statsmodels.*",
    "plotly.*",
    "boto3.*",
    "botocore.*",
]
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-v --tb=short"
asyncio_mode = "auto"
